==> Code/Notebooks/01_Scraping_Clean.ipynb <==
# Cleaned Attendance Data

#### NFL Team Attendance From 2010 - 2023. This data was scraped in conjunction with a key statistics csv file found on Kaggle to analyze how attendance trends have changed over time along with what ket statistics effect attendance numbers.#

#### The data ws scrapped from this url https://www.espn.com/nfl/attendance/_/year/20XX (with the XX depending on what year)

#### This jupiter notebook cleans the original data file of the scraped data from ESPN.com

# import required libraires

import pandas as pd

# read in the orignal data csv file
merged_data = pd.read_csv("raw_data.csv", encoding= 'utf-8')

years = [2023, 2022, 2021, 2019, 2018, 2017, 2016, 2015, 2014, 2013, 2012, 2011, 2010]

# Initialize the renaming dictionary
rename_dict = {}

# Loop through the years to create the column renaming mappings
for year in years:
    rename_dict[f'overall_total_{year}'] = f'fan_att_{year}'
    rename_dict[f'overall_avg_{year}'] = f'avg_fan_att_{year}'

# Apply the renaming
merged_data = merged_data.rename(columns=rename_dict)

# Verify the changes by displaying the updated columns
print("Updated columns after renaming:")
print(merged_data.columns)

merged_data.columns

merged_data.dtypes

merged_data.to_csv('nfl_attendance_cleaned.csv', header=True, index=False, encoding='utf-8')

long_merged_data = pd.melt(
    merged_data,
    id_vars=["team"],  
    var_name="year_metric",  
    value_name="value"  
)

long_merged_data["year"] = long_merged_data["year_metric"].str.extract(r'(\d{4})')  
long_merged_data["metric"] = long_merged_data["year_metric"].str.extract(r'(fan_att|avg_fan_att)')

final_merged_data = long_merged_data.pivot_table(
    index=["team", "year"],  
    columns="metric",  
    values="value",  
    aggfunc="first"  
).reset_index()

final_merged_data.columns.name = None
final_merged_data = final_merged_data.rename(columns={"fan_att": "Fan Attendance", "avg_fan_att": "Average Fan Attendance"})

display(final_merged_data.head())

final_merged_data.to_csv("fan_attendance.csv", index=False, encoding="utf-8")
