==> Code/Notebooks/02_Kaggle_DF.ipynb <==
# Kaggle CSV Cleaning

# Cleaned Attendance Data

#### The data csv file found from Kaggle gives us ket statistics of NFL teams.

#### Kaggle Dataset: https://www.kaggle.com/datasets/nickcantalupa/nfl-team-data-2003-2023 

#### This jupiter notebook cleans the original data file of the Kaggle Dataset. 

import pandas as pd
import matplotlib.pyplot as plt

# read the file in 
nfl_team_data = pd.read_csv("team_stats_2003_2023.csv", encoding='utf-8') 
display(nfl_team_data.head())

# Make a copy of the data 
nfl_team_data_edited = nfl_team_data.copy()

# work from the copy and display dataframe 
display(nfl_team_data_edited.head())

print(nfl_team_data_edited.dtypes)

# edit ties column to be int64 
nfl_team_data_edited['ties'] = nfl_team_data_edited['ties'].astype('Int64')

# view data types again 
print(nfl_team_data_edited.dtypes)

# check the shape of the dataframe 
nfl_team_data_edited.shape

# Check for duplicates 
duplicate_records = nfl_team_data_edited[nfl_team_data_edited.duplicated(keep=False)]

# keep=False will identify and display every row that has duplicates, ensuring that all instances of the duplicated data are shown
print(f"total records: {len(nfl_team_data_edited)} \n")
print(f"duplicate records: {len(nfl_team_data_edited[nfl_team_data_edited.duplicated(keep="first")])} \n")
print("All instances of the duplicated data","\n", "-"*50)
display(duplicate_records)

# look at the unique values of the columns 
nfl_team_data_edited['ties'].unique()

# fix Nan values in ties 
nfl_team_data_edited['ties'] = nfl_team_data_edited['ties'].fillna(0)

# check the values again
nfl_team_data_edited['ties'].unique()

nfl_team_data_edited['mov'].unique()

nfl_team_data_edited['team'].unique()

# List of columns to drop
columns_to_drop = [
    'points', 'points_opp', 'points_diff', 'mov', 'g', 'plays_offense', 
    'yds_per_play_offense', 'fumbles_lost', 'first_down', 'pass_cmp', 
    'pass_att', 'pass_yds', 'pass_td', 'pass_int', 'pass_net_yds_per_att', 
    'pass_fd', 'rush_att', 'rush_yds', 'rush_td', 'rush_yds_per_att', 
    'rush_fd', 'penalties', 'penalties_yds', 'pen_fd', 'score_pct', 
    'turnover_pct', 'exp_pts_tot', 'ties'
]

# Drop the columns from the dataframe
nfl_team_data_edited = nfl_team_data_edited.drop(columns=columns_to_drop, errors='ignore')

display(nfl_team_data_edited)

nfl_team_data_edited.to_csv('nfl_team_data_cleaned.csv', header=True, index=False, encoding='utf-8')
