==> Code/Notebooks/03_Merged.ipynb <==
# Final Merged Dataframe & Analysis

# Cleaned Attendance Data

#### This project uses scraped data from ESPN in conjunction with a key statistics csv file found on Kaggle to analyze how attendance trends have changed over time along with what ket statistics effect attendance numbers.#

#### Attendence Data: https://www.espn.com/nfl/attendance/_/year/20XX (with the XX depending on what year)
#### Statistics Data: https://www.kaggle.com/datasets/nickcantalupa/nfl-team-data-2003-2023

#### This notbooks merges the ESPN scraped data and the csv file contianing ket statistics and analyzes how the two are correlated and what realtionship they share. We also analyze attednance trends and what teams fans are loyal or not. 

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import scipy.stats as stats

attendance_data = pd.read_csv("fan_attendance.csv", encoding="utf-8")
team_data = pd.read_csv("nfl_team_data_cleaned.csv", encoding="utf-8")

# Merge the dataframes on 'team' and 'year'
merged_df = pd.merge(attendance_data, team_data, on=["team", "year"], how="inner")

display(merged_df)

merged_df.dtypes

# change the columns to be an integer

merged_df['year'] = merged_df['year'].astype(int)
merged_df['Average Fan Attendance'] = merged_df['Average Fan Attendance'].astype(str).str.replace(',', '').astype(int)
merged_df['Fan Attendance'] = merged_df['Fan Attendance'].astype(str).str.replace(',', '').astype(int)

display(merged_df)

# save the into a final csv file

merged_df.to_csv("final_df", encoding='utf-8')
